unify color handling in tomselect
