+++ assets/js/KimaiLoader.js
@@ -42,6 +42,7 @@ import KimaiNotification from "./plugins/KimaiNotification";
 import KimaiHotkeys from "./plugins/KimaiHotkeys";
 import KimaiRemoteModal from "./plugins/KimaiRemoteModal";
 import KimaiUser from "./plugins/KimaiUser";
+import KimaiAutocompleteTags from "./forms/KimaiAutocompleteTags";
 
 export default class KimaiLoader {
 
@@ -70,6 +71,7 @@ export default class KimaiLoader {
         kimai.registerPlugin(new KimaiDateRangePicker('input[data-daterangepicker="on"]'));
         kimai.registerPlugin(new KimaiDatePicker('input[data-datepicker="on"]'));
         kimai.registerPlugin(new KimaiAutocomplete());
+        kimai.registerPlugin(new KimaiAutocompleteTags());
         kimai.registerPlugin(new KimaiTimesheetForm());
         kimai.registerPlugin(new KimaiTeamForm());
         kimai.registerPlugin(new KimaiCopyDataForm());
