+++ assets/js/forms/KimaiFormSelect.js
@@ -10,9 +10,9 @@
  */
 
 import TomSelect from 'tom-select';
-import KimaiFormPlugin from "./KimaiFormPlugin";
+import KimaiFormTomselectPlugin from "./KimaiFormTomselectPlugin";
 
-export default class KimaiFormSelect extends KimaiFormPlugin {
+export default class KimaiFormSelect extends KimaiFormTomselectPlugin {
 
     constructor(selector, apiSelects)
     {
@@ -88,24 +88,13 @@ export default class KimaiFormSelect extends KimaiFormPlugin {
         };
 
         let render = {
-            option_create: (data, escape) => {
-                const name = escape(data.input);
-                if (name.length < 3) {
-                    return null;
-                }
-                const tpl = this.translate('select.search.create');
-                const tplReplaced = tpl.replace('%input%', '<strong>' + name + '</strong>');
-                return '<div class="create">' + tplReplaced + '</div>';
-            },
-            no_results: (data, escape) => {
-                const tpl = this.translate('select.search.notfound');
-                const tplReplaced = tpl.replace('%input%', '<strong>' + escape(data.input) + '</strong>');
-                return '<div class="no-results">' + tplReplaced + '</div>';
-            },
             onOptionAdd: (value) => {
                 node.dispatchEvent(new CustomEvent('create', {detail: {'value': value}}));
             },
         };
+
+        const rendererType = (node.dataset['renderer'] !== undefined) ? node.dataset['renderer'] : 'default';
+        options.render = {...render, ...this.getRenderer(rendererType)};
 
         if (node.dataset['create'] !== undefined) {
             options = {...options, ...{
@@ -122,44 +111,6 @@ export default class KimaiFormSelect extends KimaiFormPlugin {
         if (node.dataset.disableSearch !== undefined) {
             options = {...options, ...{
                 controlInput: null,
-            }};
-        }
-
-        if (node.dataset['renderer'] !== undefined && node.dataset['renderer'] === 'color') {
-            options.render = {...render, ...{
-                option: function(data, escape) {
-                    let item = '<div class="list-group-item border-0 p-1 ps-2 text-nowrap">';
-                    if (data.color !== undefined) {
-                        item += '<span style="background-color:' + data.color + '" class="color-choice-item">&nbsp;</span>';
-                    } else {
-                        item += '<span class="color-choice-item">&nbsp;</span>';
-                    }
-                    item += escape(data.text) + '</div>';
-                    return item;
-                },
-                item: function(data, escape) {
-                    let item = '<div class="text-nowrap">';
-                    if (data.color !== undefined) {
-                        item += '<span style="background-color:' + data.color + '" class="color-choice-item">&nbsp;</span>';
-                    } else {
-                        item += '<span class="color-choice-item">&nbsp;</span>';
-                    }
-                    item += escape(data.text) + '</div>';
-                    return item;
-                }
-            }};
-        } else {
-            options.render = {...render, ...{
-                // the empty entry would collapse and only show as a tiny 5px line if there is no content inside
-                option: function(data, escape) {
-                    let text = data.text;
-                    if (text === null || text.trim() === '') {
-                        text = '&nbsp;';
-                    } else {
-                        text = escape(text);
-                    }
-                    return '<div>' + text + '</div>';
-                }
             }};
         }
